+++ CHANGELOG.md
@@ -5,13 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Update ArrowInline default arrow direction to down
+- [Patch] Update Dropdown component to remove arrowIcon `true` option
 - [Patch] Update <Steps> styling to have an outline for the Active step and filled circles for Complete steps
 
 ### Fixed
 - [Patch] Fixed <Select> export by exporting it in src/main.js
 
 ## 42.2.1 - 2019-02-25
-  - [Patch] Lockdown highlight.js dependency to version 9.5.0
+- [Patch] Lockdown highlight.js dependency to version 9.5.0
 
 ## 42.2.0 - 2019-02-22
 - [Patch] Change edit pencil to face opposite direction in <EditableInput> component